UI: (GTM_org_owner) Create delete city modal
