+++ app/src/app/[lng]/organization/[id]/account-settings/project/index.tsx
@@ -1,48 +1,8 @@
 import { useTranslation } from "@/i18n/client";
-import {
-  Accordion,
-  Box,
-  BreadcrumbItem,
-  Button,
-  Flex,
-  Heading,
-  Icon,
-  IconButton,
-  Progress,
-  Spinner,
-  Table,
-  Tabs,
-  Text,
-} from "@chakra-ui/react";
-import {
-  MdAdd,
-  MdChevronRight,
-  MdMoreVert,
-  MdOutlineFolder,
-} from "react-icons/md";
-import {
-  AccordionItem,
-  AccordionItemContent,
-  AccordionItemTrigger,
-  AccordionRoot,
-} from "@/components/ui/accordion";
-import { LuChevronDown } from "react-icons/lu";
+import { Box, Heading, Text } from "@chakra-ui/react";
 import ProgressLoader from "@/components/ProgressLoader";
-import DataTableCore from "@/components/ui/data-table-core";
-import {
-  OrganizationRole,
-  ProjectUserResponse,
-  ProjectWithCities,
-} from "@/util/types";
+import { OrganizationRole } from "@/util/types";
 import { Trans } from "react-i18next";
-import { Tag } from "@/components/ui/tag";
-import {
-  MenuContent,
-  MenuItem,
-  MenuRoot,
-  MenuTrigger,
-} from "@/components/ui/menu";
-import { RiDeleteBin6Line } from "react-icons/ri";
 import React, { useEffect, useMemo, useState, useCallback } from "react";
 import {
   useGetOrganizationQuery,
@@ -51,26 +11,14 @@ import {
 } from "@/services/api";
 import { useRouter } from "next/navigation";
 import { uniqBy } from "lodash";
-import { CircleFlag } from "react-circle-flags";
-import {
-  BreadcrumbCurrentLink,
-  BreadcrumbLink,
-  BreadcrumbRoot,
-} from "@/components/ui/breadcrumb";
-import { BsDownload } from "react-icons/bs";
 import { TFunction } from "i18next";
+import ProjectList from "@/app/[lng]/organization/[id]/account-settings/project/projectList";
+import ProjectDetails from "@/app/[lng]/organization/[id]/account-settings/project/projectDetails";
 
-const TagMapping = {
+export const TagMapping = {
   [OrganizationRole.ORG_ADMIN]: { color: "green", text: "owner" },
   [OrganizationRole.ADMIN]: { color: "blue", text: "admin" },
   [OrganizationRole.COLLABORATOR]: { color: "yellow", text: "collaborator" },
-};
-
-const getInventoryLastUpdated = (lastUpdated: Date, t: Function) => {
-  if (!lastUpdated || isNaN(new Date(lastUpdated).getTime())) {
-    return <p>{t("no-date-available")}</p>;
-  }
-  return <p>{new Date(lastUpdated).toLocaleDateString()}</p>;
 };
 
 interface UseProjectDataProps {
@@ -187,581 +135,6 @@ const ProjectOverview: React.FC<ProjectOverviewProps> = ({
   </Box>
 );
 
-interface ProjectListProps {
-  t: Function;
-  projects: ProjectWithCities[];
-  selectedProjectId: string[];
-  setSelectedProject: (value: string[]) => void;
-  setSelectedCity: (value: string | null) => void;
-}
-
-const ProjectList: React.FC<ProjectListProps> = ({
-  t,
-  projects,
-  selectedProjectId,
-  setSelectedProject,
-  setSelectedCity,
-}) => (
-  <Box w="250px" flex={1}>
-    <Text
-      fontSize="title.md"
-      mb={3}
-      fontWeight="semibold"
-      color="content.secondary"
-    >
-      {t("projects")}
-    </Text>
-    <AccordionRoot
-      variant="plain"
-      value={selectedProjectId}
-      onValueChange={(val) => {
-        setSelectedProject(val.value);
-        setSelectedCity(null);
-      }}
-    >
-      {projects.map((item) => (
-        <AccordionItem key={item.projectId} value={item.projectId}>
-          <AccordionItemTrigger
-            onClick={() => setSelectedCity(null)}
-            w="full"
-            hideIndicator
-            padding="0px"
-          >
-            <Button
-              rounded={0}
-              variant="plain"
-              display="flex"
-              justifyContent="space-between"
-              w="full"
-              minH="56px"
-              p={4}
-              pr={0}
-              alignItems="center"
-              color={
-                selectedProjectId.includes(item.projectId)
-                  ? "interactive.secondary"
-                  : "content.secondary"
-              }
-            >
-              <Text
-                fontSize="label.lg"
-                fontWeight="semibold"
-                color="currentcolor"
-              >
-                {item.name}
-              </Text>
-              <Accordion.ItemIndicator
-                color="currentColor"
-                rotate={{ base: "-90deg", _open: "-180deg" }}
-              >
-                <Icon as={LuChevronDown} color="currentColor" boxSize={4} />
-              </Accordion.ItemIndicator>
-            </Button>
-          </AccordionItemTrigger>
-          <AccordionItemContent padding="0px" pb={4}>
-            {item.cities.length === 0 ? (
-              <Text fontSize="body.lg" fontWeight={600} color="content.primary">
-                {t("no-cities")}
-              </Text>
-            ) : (
-              <Tabs.Root
-                display="flex"
-                mt="12px"
-                flexDirection="row"
-                variant="subtle"
-                w="full"
-                gap="12px"
-                value={selectedProjectId[0]} // Use selectedProjectId[0] as value to indicate selected project
-                onValueChange={(val) => setSelectedCity(val.value)}
-              >
-                <Tabs.List
-                  w="full"
-                  display="flex"
-                  flexDirection="column"
-                  gap="12px"
-                >
-                  {item.cities.map((city) => (
-                    <Tabs.Trigger
-                      key={city.cityId}
-                      value={city.cityId}
-                      fontFamily="heading"
-                      justifyContent={"left"}
-                      letterSpacing={"wide"}
-                      color="content.secondary"
-                      lineHeight="20px"
-                      fontStyle="normal"
-                      fontSize="label.lg"
-                      minH="52px"
-                      w="full"
-                      _selected={{
-                        color: "content.link",
-                        fontSize: "label.lg",
-                        fontWeight: "medium",
-                        backgroundColor: "background.neutral",
-                        borderRadius: "8px",
-                        borderWidth: "1px",
-                        borderStyle: "solid",
-                        borderColor: "content.link",
-                      }}
-                    >
-                      {city.name}
-                    </Tabs.Trigger>
-                  ))}
-                </Tabs.List>
-              </Tabs.Root>
-            )}
-          </AccordionItemContent>
-        </AccordionItem>
-      ))}
-    </AccordionRoot>
-  </Box>
-);
-
-interface ProjectDetailsProps {
-  t: Function;
-  lng: string;
-  router: any;
-  selectedCity: string | null;
-  selectedProjectData: ProjectWithCities | null | undefined;
-  selectedCityData:
-    | {
-        cityId: string;
-        name: string;
-        countryLocode: string;
-        inventories: {
-          inventoryId: string;
-          year: number;
-          lastUpdated: string;
-        }[];
-      }
-    | undefined;
-  organizationName?: string;
-  projectUsers: ProjectUserResponse[] | undefined;
-  userList: ProjectUserResponse[] | undefined;
-  isLoadingProjectUsers: boolean;
-  tabValue: string;
-  setTabValue: (value: string) => void;
-  setSelectedCity: (value: string | null) => void;
-}
-
-const ProjectDetails: React.FC<ProjectDetailsProps> = ({
-  t,
-  lng,
-  router,
-  selectedCity,
-  selectedProjectData,
-  selectedCityData,
-  organizationName,
-  projectUsers,
-  userList,
-  isLoadingProjectUsers,
-  tabValue,
-  setTabValue,
-  setSelectedCity,
-}) => (
-  <Box w="full">
-    {isLoadingProjectUsers ? (
-      <ProgressLoader />
-    ) : (
-      <Box className="bg-white" p={6} rounded={2} mt={12}>
-        <Flex justifyContent="space-between" alignItems="center" mb={6}>
-          {selectedCity ? (
-            <Box>
-              <BreadcrumbRoot
-                gap="8px"
-                fontFamily="heading"
-                fontWeight="bold"
-                letterSpacing="widest"
-                fontSize="14px"
-                textTransform="uppercase"
-                separator={
-                  <Icon
-                    as={MdChevronRight}
-                    boxSize={4}
-                    color="content.primary"
-                    h="32px"
-                  />
-                }
-              >
-                <BreadcrumbItem>
-                  <BreadcrumbLink
-                    href={`/${lng}/admin`}
-                    color="content.tertiary"
-                    fontWeight="normal"
-                    truncate
-                    className="capitalize"
-                  >
-                    {selectedProjectData?.name}
-                  </BreadcrumbLink>
-                </BreadcrumbItem>
-                <BreadcrumbCurrentLink color="content.link">
-                  <Text truncate lineClamp={1} className="capitalize">
-                    {selectedCityData?.name}
-                  </Text>
-                </BreadcrumbCurrentLink>
-              </BreadcrumbRoot>
-              <Flex mt={2} gap={2}>
-                <CircleFlag
-                  countryCode={
-                    selectedCityData?.countryLocode
-                      ?.substring(0, 2)
-                      .toLowerCase() || ""
-                  }
-                  width={32}
-                />
-                <Text fontWeight="bold" fontSize="title.md" mb={2}>
-                  {selectedCityData?.name}
-                </Text>
-              </Flex>
-            </Box>
-          ) : (
-            <Text fontWeight="bold" fontSize="title.md" mb={2}>
-              {organizationName}
-            </Text>
-          )}
-          <Button
-            onClick={() => router.push(`/${lng}/onboarding/setup`)}
-            variant="outline"
-            ml="auto"
-            h="48px"
-            mt="auto"
-          >
-            <Icon as={MdAdd} h={8} w={8} />
-            {t("add-city")}
-          </Button>
-        </Flex>
-        <Tabs.Root
-          value={tabValue}
-          onValueChange={(val) => setTabValue(val.value)}
-          defaultValue="city"
-          variant="enclosed"
-        >
-          <Tabs.List p={0} w="full">
-            {selectedCityData ? (
-              <Tabs.Trigger
-                value="inventories"
-                _selected={{
-                  borderColor: "content.link",
-                  borderBottomWidth: "2px",
-                  boxShadow: "none",
-                  fontWeight: "bold",
-                  borderRadius: "0",
-                  color: "content.link",
-                }}
-              >
-                <Text fontSize="title.md" fontStyle="normal" lineHeight="24px">
-                  {t("all-inventories", {
-                    count: selectedCityData?.inventories?.length,
-                  })}
-                </Text>
-              </Tabs.Trigger>
-            ) : (
-              <Tabs.Trigger
-                value="city"
-                _selected={{
-                  borderColor: "content.link",
-                  borderBottomWidth: "2px",
-                  boxShadow: "none",
-                  fontWeight: "bold",
-                  borderRadius: "0",
-                  color: "content.link",
-                }}
-              >
-                <Text fontSize="title.md" fontStyle="normal" lineHeight="24px">
-                  {t("all-cities", {
-                    count: selectedProjectData?.cities?.length,
-                  })}
-                </Text>
-              </Tabs.Trigger>
-            )}
-            <Tabs.Trigger
-              value="collaborators"
-              _selected={{
-                borderColor: "content.link",
-                borderBottomWidth: "2px",
-                boxShadow: "none",
-                fontWeight: "bold",
-                borderRadius: "0",
-                color: "content.link",
-              }}
-            >
-              <Text fontSize="title.md" fontStyle="normal" lineHeight="24px">
-                {t("all-collaborators", { count: projectUsers?.length })}
-              </Text>
-            </Tabs.Trigger>
-          </Tabs.List>
-          <Tabs.Content value="city">
-            <DataTableCore
-              data={selectedProjectData?.cities ?? []}
-              columns={[
-                { header: t("name"), accessor: "name" },
-                { header: t("inventories"), accessor: null },
-                { header: "", accessor: null },
-              ]}
-              renderRow={(item, idx) => (
-                <Table.Row key={idx}>
-                  <Table.Cell>
-                    <Button
-                      variant="ghost"
-                      color="content.primary"
-                      onClick={() => setSelectedCity(item.cityId)}
-                    >
-                      <Text
-                        color="content.link"
-                        fontWeight="normal"
-                        className="truncate capitalize underline"
-                        fontSize="label.lg"
-                      >
-                        {item.name}
-                      </Text>
-                    </Button>
-                  </Table.Cell>
-                  <Table.Cell>{item.inventories.length}</Table.Cell>
-                  <Table.Cell w={10} className="w-10">
-                    <MenuRoot>
-                      <MenuTrigger>
-                        <IconButton
-                          data-testid="activity-more-icon"
-                          aria-label="more-icon"
-                          variant="ghost"
-                          color="content.tertiary"
-                        >
-                          <Icon as={MdMoreVert} size="lg" />
-                        </IconButton>
-                      </MenuTrigger>
-                      <MenuContent
-                        w="auto"
-                        borderRadius="8px"
-                        shadow="2dp"
-                        px="0"
-                      >
-                        <MenuItem
-                          value={t("delete-city")}
-                          valueText={t("delete-city")}
-                          p="16px"
-                          display="flex"
-                          alignItems="center"
-                          gap="16px"
-                          _hover={{ bg: "content.link", cursor: "pointer" }}
-                          className="group"
-                          onClick={() => {}}
-                        >
-                          <Icon
-                            className="group-hover:text-white"
-                            color="sentiment.negativeDefault"
-                            as={RiDeleteBin6Line}
-                            h="24px"
-                            w="24px"
-                          />
-                          <Text
-                            className="group-hover:text-white"
-                            color="content.primary"
-                          >
-                            {t("delete-city")}
-                          </Text>
-                        </MenuItem>
-                      </MenuContent>
-                    </MenuRoot>
-                  </Table.Cell>
-                </Table.Row>
-              )}
-            />
-          </Tabs.Content>
-          <Tabs.Content value="collaborators">
-            <DataTableCore
-              data={userList ?? []}
-              columns={[
-                { header: t("email"), accessor: "email" },
-                { header: t("role"), accessor: "role" },
-                { header: "", accessor: null },
-                { header: "", accessor: null },
-              ]}
-              renderRow={(item, idx) => (
-                <Table.Row key={idx}>
-                  <Table.Cell>{item.email}</Table.Cell>
-                  <Table.Cell title={item.role}>
-                    <Tag
-                      size="lg"
-                      colorPalette={
-                        TagMapping[item.role as OrganizationRole].color
-                      }
-                    >
-                      {TagMapping[item.role as OrganizationRole].text}
-                    </Tag>
-                  </Table.Cell>
-                  <Table.Cell className="w-10">
-                    <MenuRoot>
-                      <MenuTrigger>
-                        <IconButton
-                          data-testid="activity-more-icon"
-                          aria-label="more-icon"
-                          variant="ghost"
-                          color="content.tertiary"
-                        >
-                          <Icon as={MdMoreVert} size="lg" />
-                        </IconButton>
-                      </MenuTrigger>
-                      <MenuContent
-                        w="auto"
-                        borderRadius="8px"
-                        shadow="2dp"
-                        px="0"
-                      >
-                        <MenuItem
-                          value={t("remove-user")}
-                          valueText={t("remove-user")}
-                          p="16px"
-                          display="flex"
-                          alignItems="center"
-                          gap="16px"
-                          _hover={{ bg: "content.link", cursor: "pointer" }}
-                          className="group"
-                          onClick={() => {}}
-                        >
-                          <Icon
-                            className="group-hover:text-white"
-                            color="sentiment.negativeDefault"
-                            as={RiDeleteBin6Line}
-                            h="24px"
-                            w="24px"
-                          />
-                          <Text
-                            className="group-hover:text-white"
-                            color="content.primary"
-                          >
-                            {t("remove-user")}
-                          </Text>
-                        </MenuItem>
-                      </MenuContent>
-                    </MenuRoot>
-                  </Table.Cell>
-                </Table.Row>
-              )}
-            />
-          </Tabs.Content>
-          <Tabs.Content value="inventories">
-            <Text
-              color="content.tertiary"
-              mb={2}
-              mt={6}
-              className="uppercase"
-              fontWeight="bold"
-            >
-              {t("all-inventory-years")}
-            </Text>
-            <DataTableCore
-              data={selectedCityData?.inventories ?? []}
-              columns={[
-                { header: t("year"), accessor: "year" },
-                { header: t("status"), accessor: null },
-                { header: t("last-updated"), accessor: "lastUpdated" },
-                { header: "", accessor: null },
-                { header: "", accessor: null },
-              ]}
-              renderRow={(item, idx) => (
-                <Table.Row key={idx}>
-                  <Table.Cell>
-                    <Flex gap={2} alignItems="center">
-                      <Icon
-                        as={MdOutlineFolder}
-                        color="content.tertiary"
-                        size="lg"
-                      />
-                      <Text
-                        color="content.link"
-                        fontWeight="normal"
-                        className="truncate capitalize underline"
-                        fontSize="label.lg"
-                      >
-                        {item.year}
-                      </Text>
-                    </Flex>
-                  </Table.Cell>
-                  <Table.Cell>
-                    <Progress.Root
-                      maxW="137px"
-                      value={0}
-                      borderRadius="8px"
-                      colorScheme="baseStyle"
-                      height="8px"
-                      width="137px"
-                    >
-                      <Progress.Track>
-                        <Progress.Range />
-                      </Progress.Track>
-                    </Progress.Root>
-                  </Table.Cell>
-                  <Table.Cell>
-                    {getInventoryLastUpdated(new Date(item.lastUpdated), t)}
-                  </Table.Cell>
-                  <Table.Cell>
-                    <IconButton
-                      data-testid="download-inventory-icon"
-                      aria-label="more-icon"
-                      variant="ghost"
-                      color="content.tertiary"
-                    >
-                      <Icon as={BsDownload} size="lg" />
-                    </IconButton>
-                  </Table.Cell>
-                  <Table.Cell className="w-10">
-                    <MenuRoot>
-                      <MenuTrigger>
-                        <IconButton
-                          data-testid="activity-more-icon"
-                          aria-label="more-icon"
-                          variant="ghost"
-                          color="content.tertiary"
-                        >
-                          <Icon as={MdMoreVert} size="lg" />
-                        </IconButton>
-                      </MenuTrigger>
-                      <MenuContent
-                        w="auto"
-                        borderRadius="8px"
-                        shadow="2dp"
-                        px="0"
-                      >
-                        <MenuItem
-                          value={t("delete-inventory")}
-                          valueText={t("delete-inventory")}
-                          p="16px"
-                          display="flex"
-                          alignItems="center"
-                          gap="16px"
-                          _hover={{ bg: "content.link", cursor: "pointer" }}
-                          className="group"
-                          onClick={() => {}}
-                        >
-                          <Icon
-                            className="group-hover:text-white"
-                            color="sentiment.negativeDefault"
-                            as={RiDeleteBin6Line}
-                            h="24px"
-                            w="24px"
-                          />
-                          <Text
-                            className="group-hover:text-white"
-                            color="content.primary"
-                          >
-                            {t("delete-inventory")}
-                          </Text>
-                        </MenuItem>
-                      </MenuContent>
-                    </MenuRoot>
-                  </Table.Cell>
-                </Table.Row>
-              )}
-            />
-          </Tabs.Content>
-        </Tabs.Root>
-      </Box>
-    )}
-  </Box>
-);
-
-// --- Main Component ---
-
 const ProjectSettings = ({ lng, id }: { lng: string; id: string }) => {
   const { t } = useTranslation(lng, "settings");
   const router = useRouter();
@@ -841,6 +214,7 @@ const ProjectSettings = ({ lng, id }: { lng: string; id: string }) => {
             selectedProjectId={selectedProjectId}
             setSelectedProject={setSelectedProjectId}
             setSelectedCity={setSelectedCityId}
+            selectedCity={selectedCityId}
           />
         )}
         <ProjectDetails
