+++ app/src/i18n/locales/en/settings.json
@@ -123,5 +123,9 @@
   "all-cities": "All Cities ({{count}})",
   "all-collaborators": "All Collaborators ({{count}})",
   "all-inventories": "Inventories ({{count}})",
-  "delete-city": "Delete City"
+  "delete-city": "Delete City",
+  "confirm-city-delete": "Are you sure you want to permanently delete <strong>{{name}}</strong>",
+  "delete-all-inventories-warning": "All the inventories associated with this city will be deleted as well.",
+  "yes-i-understand": "Yes, I understand",
+  "enter-city-name-confirmation": "Please write \"{{name}}\" below to delete it"
 }
