+++ app/src/services/api.ts
@@ -48,11 +48,10 @@ import {
   ACTION_TYPES,
   ThemeResponse,
   OrganizationWithThemeResponse,
-  UpdateUserPayload
+  UpdateUserPayload,
 } from "@/util/types";
 import type { GeoJSON } from "geojson";
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
-
 
 export const api = createApi({
   reducerPath: "api",
@@ -432,10 +431,7 @@ export const api = createApi({
         providesTags: ["UserData"],
       }),
 
-      setCurrentUserData: builder.mutation<
-        UserAttributes,
-        UpdateUserPayload
-      >({
+      setCurrentUserData: builder.mutation<UserAttributes, UpdateUserPayload>({
         query: (data) => ({
           url: `/user/${data.userId}`,
           method: "PATCH",
@@ -1163,6 +1159,20 @@ export const api = createApi({
           response,
         providesTags: ["Organization"],
       }),
+      deleteCity: builder.mutation({
+        query: (cityId: string) => ({
+          method: "DELETE",
+          url: `/city/${cityId}`,
+        }),
+        transformResponse: (response: { data: any }) => response.data,
+        invalidatesTags: [
+          "CityData",
+          "Projects",
+          "CitiesAndInventories",
+          "UserData",
+          "UserInfo",
+        ],
+      }),
     };
   },
 });
@@ -1262,5 +1272,6 @@ export const {
   useGetThemesQuery,
   useSetOrgWhiteLabelMutation,
   useGetOrganizationForInventoryQuery,
+  useDeleteCityMutation,
 } = api;
 export const { useGetOCCityQuery, useGetOCCityDataQuery } = openclimateAPI;
